requirements.txt
numpy
scipy
matplotlib
h5py
sklearn
umap-learn
opencv-python
tk
Pillow